% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss.r
\name{diss}
\alias{diss}
\title{Transform a distance object to Ypma dissimilarity}
\usage{
diss(d, allsteps = FALSE)
}
\arguments{
\item{d}{An object inheriting from `dist`.}

\item{allsteps}{Logical; return the minimum spanning tree and graph as well?}
}
\value{
A square matrix of Ypma hop dissimilarities, or a list of intermediate
  objects when `allsteps = TRUE`.
}
\description{
Converts a `dist` object (including `cluster::daisy()` dissimilarities) into
minimum-spanning-tree hop distances.
}
