% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{k_pneumoniae}
\alias{k_pneumoniae}
\title{Surveillance dataset of Klebsiella pneumoniae}
\format{A data frame with 245 rows and 21 variables:
\describe{
  \item{id}{strain identifier (generated during data export, not identical to original strain id)}
  \item{time}{time of request (altered by addition of a constant value)}
  \item{unit}{ward name (remapped to fake names)}
  \item{Amikazin}{susceptibility category of antibiotic 'Amikazin'}
  \item{Ampicillin}{susceptibility category of antibiotic 'Ampicillin'}
  \item{Ampicillin+Sulbactam}{susceptibility category of antibiotic 'Ampicillin+Sulbactam'}
  \item{Cefepim}{susceptibility category of antibiotic 'Cefepim'}
  \item{Cefotaxim}{susceptibility category of antibiotic 'Cefotaxim'}
  \item{Cefpodoxim}{susceptibility category of antibiotic 'Cefpodoxim'}
  \item{Ceftazidime}{susceptibility category of antibiotic 'Ceftazidime'}
  \item{Cefuroxim}{susceptibility category of antibiotic 'Cefuroxim'}
  \item{Ciprofloxacin}{susceptibility category of antibiotic 'Ciprofloxacin'}
  \item{Co-trimoxazol}{susceptibility category of antibiotic 'Co-trimoxazol'}
  \item{Ertapenem}{susceptibility category of antibiotic 'Ertapenem'}
  \item{Fosfomycin}{susceptibility category of antibiotic 'Fosfomycin'}
  \item{Gentamicin}{susceptibility category of antibiotic 'Gentamicin'}
  \item{Imipenem}{susceptibility category of antibiotic 'Imipenem'}
  \item{Meropenem}{susceptibility category of antibiotic 'Meropenem'}
  \item{Piperacillin}{susceptibility category of antibiotic 'Piperacillin'}
  \item{Piperacillin+Tazobactam}{susceptibility category of antibiotic 'Piperacillin+Tazobactam'}
  \item{Tobramycin}{susceptibility category of antibiotic 'Tobramycin'}
}}
\usage{
k_pneumoniae
}
\description{
A surveillance dataset of susceotibility tests from three hospitals.
Columns 'id', 'time' and 'unit' have been altered to prevent identification of cases.
Susceptibility categorization (1 = susceptible, 2 = intermediate, 3 = resistant) has been performed according to EUCAST 9.0.
}
\keyword{datasets}
