% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mst}
\alias{mst}
\title{Minimum spanning tree}
\usage{
mst(d, debug = FALSE)
}
\arguments{
\item{d}{An object of class `dist` containing finite distances.}

\item{debug}{Logical; print intermediate algorithm state.}
}
\value{
A symmetric adjacency matrix of class `mst`.
}
\description{
C++ implementation based on `ape::mst`.
}
