% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mstC}
\alias{mstC}
\title{Minimum spanning tree}
\usage{
mstC(d)
}
\arguments{
\item{d}{An object of type 'dist'.}
}
\value{
An adjacency matrix belonging to class 'mst'.
}
\description{
This is a C++ transcript of function \code{mst} in package \code{ape} by Yvonnick Noel, Julien Claude, and Emmanuel Paradis, 
cf. \href(https://github.com/cran/ape/blob/master/R/mst.R){Github archive of mst.R}.
}
