% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_expand.r
\name{mx.expand}
\alias{mx.expand}
\title{Expand a square dissimilarity matrix}
\usage{
mx.expand(
  conversion.table,
  mx,
  noise0 = TRUE,
  noise.lo = 0,
  noise.hi = 0.01,
  ...
)
}
\arguments{
\item{conversion.table}{A data.frame containing two columns; the first with
id labels and the second with row/column names in `mx`.}

\item{mx}{A square numeric matrix with identical row and column names.}

\item{noise0}{Logical; add noise to off-diagonal zero values using
[zero.noise()]? Defaults to `TRUE`.}

\item{noise.lo, noise.hi}{Numeric bounds passed to [zero.noise()].}

\item{...}{Further arguments passed to [zero.noise()].}
}
\value{
A matrix with ids from `conversion.table` as row/column names.
}
\description{
This function accepts a conversion table and a square matrix as input to
return a new square matrix with row and column names referring to the ids in
the conversion table.
}
