% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{s_aureus}
\alias{s_aureus}
\title{Surveillance dataset of Staphylococcus aureus}
\format{A data frame with 474 rows and 25 variables:
\describe{
  \item{id}{strain identifier (generated during data export, not identical to original strain id)}
  \item{time}{time of request (altered by addition of a constant value)}
  \item{unit}{ward name (remapped to fake names)}
  \item{Ampicillin}{susceptibility category of antibiotic 'Ampicillin'}
  \item{Ampicillin+Sulbactam}{susceptibility category of antibiotic 'Ampicillin+Sulbactam'}
  \item{Cefpodoxim}{susceptibility category of antibiotic 'Cefpodoxim'}
  \item{Cefuroxim}{susceptibility category of antibiotic 'Cefuroxim'}
  \item{Ciprofloxacin}{susceptibility category of antibiotic 'Ciprofloxacin'}
  \item{Clindamycin}{susceptibility category of antibiotic 'Clindamycin'}
  \item{Co-trimoxazol}{susceptibility category of antibiotic 'Co-trimoxazol'}
  \item{Daptomycin}{susceptibility category of antibiotic 'Daptomycin'}
  \item{Erythromycin}{susceptibility category of antibiotic 'Erythromycin'}
  \item{Fosfomycin}{susceptibility category of antibiotic 'Fosfomycin'}
  \item{Gentamicin}{susceptibility category of antibiotic 'Gentamicin'}
  \item{Imipenem}{susceptibility category of antibiotic 'Imipenem'}
  \item{Linezolid}{susceptibility category of antibiotic 'Linezolid'}
  \item{Moxifloxacin}{susceptibility category of antibiotic 'Moxifloxacin'}
  \item{Oxacillin}{susceptibility category of antibiotic 'Oxacillin'}
  \item{Penicillin G}{susceptibility category of antibiotic 'Penicillin G' (Benzylpenicillin)}
  \item{Piperacillin+Tazobactam}{susceptibility category of antibiotic 'Piperacillin+Tazobactam'}
  \item{Rifampicin}{susceptibility category of antibiotic 'Rifampicin'}
  \item{Teicoplanin}{susceptibility category of antibiotic 'Teicoplanin'}
  \item{Tetrazyklin}{susceptibility category of antibiotic 'Tetrazyklin' (tetracyline)}
  \item{Tigecyclin}{susceptibility category of antibiotic 'Tigecyclin' (tigecycline)}
  \item{Vancomycin}{susceptibility category of antibiotic 'Vancomycin'}
}}
\usage{
s_aureus
}
\description{
A surveillance dataset of susceotibility tests from three hospitals.
Columns 'id', 'time' and 'unit' have been altered to prevent identification of cases.
Susceptibility categorization (1 = susceptible, 2 = intermediate, 3 = resistant) has been performed according to EUCAST 9.0, with the exception of Rifampicin.
}
\keyword{datasets}
