% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{units_effdist}
\alias{units_effdist}
\title{Matrix of effective distances between wards.}
\format{A quadratic matrix with 78 rows and 78 columns.}
\usage{
units_effdist
}
\description{
A quadratic matrix containing effective distances between wards. Effective distances have been
calculated from \code{\link{igraph}} object \code{\link{units_igraph}} using function \code{\link{graph2effdist}}, which depends on package \code{\link{NetOrigin}}.
Ward names (column and row ids) have been remapped to fake names to prevent their identification.
}
\keyword{datasets}
