% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{units_igraph}
\alias{units_igraph}
\title{Network graph of 78 units of a hospital trust consisting of three hospitals}
\format{An \code{\link{igraph}} object}
\usage{
units_igraph
}
\description{
An \code{\link{igraph}} object with 78 nodes representing units (mostly wards) generated from laboratory request data (full blood count) over a month.
Movements across nodes are assumed to represent patient movements.
}
\keyword{datasets}
