% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_clust.r
\name{y.clust}
\alias{y.clust}
\title{Identify putative clusters}
\usage{
y.clust(yd, hs = c(2, 4, 6), result = c("sizes", "clusters"))
}
\arguments{
\item{yd}{A square Ypma dissimilarity matrix or a list of such matrices.}
\item{hs}{Numeric heights at which hierarchical trees should be cut.}
\item{result}{Return cluster-size thresholds (\code{"sizes"}) or cluster details (\code{"clusters"}).}
}
\value{
A data.frame describing clusters or permutation thresholds.
}
\description{
Identify putative clusters.
}
