% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_noise.r
\name{zero.noise}
\alias{zero.noise}
\title{Add random noise to zero values in a square matrix}
\usage{
zero.noise(x, lo = 0, hi = 0.01, exclude.diag = TRUE)
}
\arguments{
\item{x}{A numeric square matrix.}

\item{lo, hi}{Numeric lower and upper bounds for uniform random values.}

\item{exclude.diag}{Logical; keep diagonal values at zero?}
}
\value{
A numeric square matrix with noise replacing zero values.
}
\description{
Add random noise to zero values in a square matrix
}
