% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_search.r
\name{cluster.search}
\alias{cluster.search}
\title{Search for putative clusters}
\usage{
cluster.search(d, e, tc, ic, uc, p.n, hs, ...)
}
\arguments{
\item{d}{A data.frame containing id, time, unit and type columns.}
\item{e}{A square effective-distance matrix for units in \code{d}.}
\item{tc, uc, ic}{Character names or numeric positions of time, unit and id.}
\item{p.n}{Number of permutations.}
\item{hs}{Heights at which hierarchical trees are cut.}
\item{...}{Arguments passed to \code{ypma.diss()}.}
}
\value{
A list of data.frames, one for each significant putative cluster.
}
\description{
Searches surveillance observations for clusters with unusually small combined spatial, type/genetic, and temporal Ypma dissimilarity.
}
