% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss.r
\name{diss}
\alias{diss}
\title{Transform distance or dissimilarity object to Ypma dissimilarity}
\usage{
diss(d, allsteps = FALSE)
}
\arguments{
\item{d}{An object of type type \code{dist} or type \code{dissimilarity}.}

\item{allsteps}{A logical specifying if intermediate steps (minimum spanning tree, graph) are to be returned. Defaults to \code{FALSE}.}
}
\value{
A square matrix containing Ypma dissimilarities.
}
\description{
This function transforms objects of type \code{dist} (output from \code{stats::dist})
or \code{dissimilarity} (output from \code{cluster::daisy}) into an Ypma dissimilarity matrix.
}
