% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph2effdist.r
\name{graph2effdist}
\alias{graph2effdist}
\title{Calculate effective distances between nodes of a graph}
\usage{
graph2effdist(g, edge.attribute = NULL)
}
\arguments{
\item{g}{An igraph object.}

\item{edge.attribute}{An attribute from the graph, e.g. weight.}
}
\value{
A square matrix with effective distances between nodes.
}
\description{
This function accepts an igraph object generated by package \code{\link{igraph}} and calcuates the effective distances.
}
\examples{
# generates object 'units_effdist':
graph2effdist(units_igraph, edge.attribute = "weight")
}
