% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_expand.r
\name{mx.expand}
\alias{mx.expand}
\title{Expand a square dissimilarity matrix}
\usage{
mx.expand(
  conversion.table,
  mx,
  noise0 = TRUE,
  noise.lo = 0,
  noise.hi = 0.01,
  ...
)
}
\arguments{
\item{conversion.table}{A data.frame containting two columns; the first with id labels, the second with the names of rows (= names of columns) in square matrix \code{mx} (see below), to which it refers.}

\item{mx}{A square numerical matrix with equal row and column names that the second column of \code{conversion.table} refers to.}

\item{noise0}{A logical determining whether to add noise to zero values using function \code{zero.noise}. Defaults to TRUE.}

\item{noise.lo}{A numerical passed to \code{zero.noise}, specifying the lower boundary of the random value to be inserted. Only used if \code{noise0} is TRUE.}

\item{noise.hi}{A numerical passed to \code{zero.noise}, specifying the upper boundary of the random value to be inserted. Only used if \code{noise0} is TRUE.}

\item{...}{Further arguments passed to \code{zero.noise}.}
}
\value{
A matrix with id in \code{conversion.table} used as row/column names.
}
\description{
This function accepts a conversion table and a square matrix as input to return a new square matrix with changed row and column names referring to id labels in conversion table.
}
