% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_clust.r
\name{y.clust}
\alias{y.clust}
\title{Identify putative clusters}
\usage{
y.clust(yd, hs = c(2, 4, 6), result = c("sizes", "clusters"))
}
\arguments{
\item{yd}{A square matrix (Ypma dissimilarity matrix) or a list thereof.}

\item{hs}{An integer containing the heights at wich trees from hierarchical clustering should be cut.}

\item{return}{An character specifying if sizes of putative clusters with according smallest maximal dissimilarities ('sizes') or clusters with information on ids etc. ('clusters') are to be returned.}
}
\value{
A square matrix containing Ypma dissimilarities.
}
\description{
This function accepts a Ypma dissimilarity matrix (or a list thereof) and returns information on putative clusters.
}
