% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_noise.r
\name{zero.noise}
\alias{zero.noise}
\title{Add random noise to zero values in square matrix}
\usage{
zero.noise(x, lo = 0, hi = 0.01, exclude.diag = TRUE)
}
\arguments{
\item{x}{A numerical square matrix.}

\item{lo}{A numeric specifying the lower boundary of to be inserted random values (random values are generated with \code{\link{runif}}).}

\item{hi}{A numeric specifying the upper boundary of to be inserted random values.}

\item{exclude.diag}{A logical stating whether the diagonal values should be left alone; defaults to TRUE.}
}
\value{
A numerical square matrix with random noise added to zero values.
}
\description{
This is a helper function used by \code{\link{mx.expand}}.
}
